requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
jinja2==3.1.4
python-dotenv==1.0.1
sqlmodel==0.0.22
openai==1.58.1
streamlit==1.40.1

